refactor(web): fold cross-row ghost label helper into calendarInteractionDom
crossRowDraftEventDom.ts existed only because its author and the file's owner
were editing concurrently during the same work session - a one-function file
whose function belongs beside the updateCalendarDraftEventTimeLabel it
complements. Fold it in; no behavior change.
